Move factory function to util file
